renamed using kebab case...
Closes #666
